extract HitTable component from HitGrid
