+++ ui/src/components/elements/hit/grid/EnhancedCell.tsx
