+++ ui/src/components/routes/hits/search/SearchPane.tsx
@@ -16,6 +16,7 @@ import SearchTotal from 'components/elements/addons/search/SearchTotal';
 import HowlerCard from 'components/elements/display/HowlerCard';
 import HitBanner from 'components/elements/hit/HitBanner';
 import HitCard from 'components/elements/hit/HitCard';
+import HitContextMenu from 'components/elements/hit/HitContextMenu';
 import { HitLayout } from 'components/elements/hit/HitLayout';
 import useHitSelection from 'components/hooks/useHitSelection';
 import { useMyLocalStorageItem } from 'components/hooks/useMyLocalStorage';
@@ -28,7 +29,6 @@ import { useLocation, useParams } from 'react-router-dom';
 import { useContextSelector } from 'use-context-selector';
 import { StorageKey } from 'utils/constants';
 import { BundleScroller } from './BundleScroller';
-import HitContextMenu from './HitContextMenu';
 import HitQuery from './HitQuery';
 import QuerySettings from './QuerySettings';
 import SearchActionMenu from './shared/SearchActionMenu';
