+++ ui/src/components/routes/home/ViewCard.tsx
@@ -5,9 +5,9 @@ import AppListEmpty from 'commons/components/display/AppListEmpty';
 import { useHitContextSelector } from 'components/app/providers/HitProvider';
 import { ViewContext } from 'components/app/providers/ViewProvider';
 import HitBanner from 'components/elements/hit/HitBanner';
+import HitContextMenu from 'components/elements/hit/HitContextMenu';
 import { HitLayout } from 'components/elements/hit/HitLayout';
 import useMyApi from 'components/hooks/useMyApi';
-import HitContextMenu from 'components/routes/hits/search/HitContextMenu';
 import type { Hit } from 'models/entities/generated/Hit';
 import { useCallback, useEffect, useMemo, useRef, useState, type FC } from 'react';
 import { useTranslation } from 'react-i18next';
